Add BERT-based type classification model
Add methods to prepare and reshape the Wikidata-TypeRec dataset.
Add BERT-based model with train() and test() methods. It is trained to
predict a high-level entity type based on a mention with its context.
The mention and the context are concatenated with the [SEP] token for
the input of BERT. Only positive samples are used, as it is a
multi-class classification task, not a binary classification task.

For the training, a data generator method is used as wrapper for the
dataset. For testing the dataset is used directly.

The validation dataset is currently identical with the training
dataset.

For metrics, precision, recall and F1 are calculated for each epoch and
also printed in average for all epochs.

Extract list of high-level types to new module typerec.types.