Integrate type classification model into EL system
Add BERTTypeFilter as new filter based on the BERT-based TypeRec model
for type classification.
Rename spaCy-based filter to NERTypeFilter.

Filter is now an distinct step of the entity linking pipeline and not
integrated into candidate generation anymore.

Create subclasses ELConfig and ModelConfig for entity linking and
model training respectively.
Move directory definitions to config module.

Extract methods for handling types and supertypes to the types module.

Add internal method __initialize() to the ELModel s.t. it is only
initialized when it is trained, but not anymore when it merely is
loaded from a checkpoint for testing or predicting (hence no training
config is required then).