Refactor Wikidata-TypeRec dataset creation
Merge code for generation and analysis of Wikidata-TypeRec dataset.
Remove old modules. Move code into typerec module.
Remove negative samples as well as the samples based on the wrong
entities from Cetoli.
Add neutral type for entities that don't match the current list of
supertypes.
Add TypeRecDatasetConfig for the creation of the dataset.