Make filter ignore default type
When predicting the default type for a mention, its candidates are not
filtered now anymore. Reason: Otherwise all candidates are removed, as
no Wikidata type is a subclass of the default type, which is a manually
chosen string (currently "OTHER").
Add the setting attribute "filter_default_type" to configure this
behavior.
Add a document with a mention for testing this.