Fix static matching score
The predicted matching score of the candidate entites was overwritten
and fixed to 1.0, i.e. ignored completely for chosing the best
candidate.
Check if all candidate entities were removed in the disambiguation
(e.g. because they don't have an PBG embedding) before trying to pick
a best to avoid exception.
Add package sys to inout.wikidata.